skill gap analyser/requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
youtube-transcript-api